Separate unit from integration tests for ReusableChunk and remove unnecessary SFBaseTest parent for unit test classes
